refactor: UserDetails의 구현부 변경 및 관련 코드 수정
